+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Added explanatory text to CivisML_parallel_training.ipynb (#126).
 
+### Fixed
+- Added `ResourceWarning` for Python 2.7 (#128).
+
 ### Added
 - Jupyter notebook with demonstrations of use patterns and abstractions in the Python API client (#127).